requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
dotenv-python==0.0.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.9
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
